implement new Web Interfaces
